Changes for t38modem to work with Opal 3.10.11
Includes a Busy Fix -- When we are busy we now set the cause code
to EndedByLocalBusy instead of EndedByLocalUser.

Changed version to 3.11.0. The first number in the version will
track the second number in Opal, where 3 -> 10 and 4 -> 12.  The
second number in version will track the third number in Opal exactly.

Also changed Makefile back to the pkgconfig version.