Fix missing DTMF media line in SDP on Invite.
Needed to add to FormatsMask in opal/manager (formats we want to keep)
both UserInput/RFC2833 and NamedSignalEvent. This missing info
in SDP kept Alcatel systems from allowing us to make a phone call.

We are also bumping the version number to v3.13.0 to pickup opal
v3.10.13 which has a Cisco CUBE bug work-around.