Version 3.14.0
Added new option --sip-retry-403-forbidden
This option is needed because UNIFY systems will respond 403 Forbiden
when resetting to apply new settings. RFC 3261 syas this is a final
answer and the request should not be retried. UNIFY syas that is the
way they work and you have to retry. This violates the RFC, but
is needed to make UNIFY work.

Added the ability to use all of the Fake Codecs:
  G.722, G.722.1, G.722.2, G.726-40K, G.726-32K, G.726-24K, G.726-16K,
  G.728, G.729, G.729A, G.729B, G.729A/B, G.723.1, G.723.1(5.3k),
  G.723.1A(6.3k), G.723.1A(5.3k), GSM-06.10, GSM-AMR, iLBC
Specify with --fake-audio and --sip-audio or --h323-audio.
Force fax mode may need to be used because to actual sound is
transmitted with the fake codecs.