Add FastFoIP feature
FastFoIP is enabled via --fast-t38-mode or
--fast-t38-mode-regex. When using --fast-t38-mode all calls
made will use FastFoIP. When using --fast-t38-mode-regex then
only calls made to numbers that match the regex will use FastFoIP.
FastFoIP eliminates some of the delays to speed up the fax call.
It should only be used when it it known that there is no analog
portion of the call. If there is an analog portion, the call
will fail.