Fix T38Modem hang when sending non-ECM data and the line drops.
Problem found at Regions Bank.

This is a rather serious problem where a line drop during a non-ECM send fax
will cause T38Modem to stop receiving data on the pty. The problem was introduced
on 2012-11-09 with commit 728b05 "Added compatibility for OPAL >= 3.10.5".
In this commit the code was changed to use InternalClose() instead of Close() in
in T38ModemMediaStream. The check for isOpen was left in, but Opal has already
turned off isOpen before calling InternalClose. Therefore, the code inside that
if, the call to t38engine->CloseOut() is not run and that causes the backup of
data that causes the hang.