Restore offer of OpalPCM16. All types of fax transmissions are working.
With this level of code all types of fax transmission are working,

options: -X 3 -e
  T.38 --> T.38
  T.38 --> G.711 (fallback)
  T.38 <-- T.38
  T.38 <-- G.711 (fallback)

options: -a
  G.711 --> T.38 (fallback)
  G.711 --> G.711
  G.711 <-- T.38 (fallback)
  G.711 <-- G.711

When T.38 is enabled, we start every call with PCM16<>G.711.
the switch to T.38 can be triggered by:
  On Receive Calls:
    After 1 second
    Receiving CNG if -e option is used
  On Send Calls:
    Receiving CED
    On Timeout if -X n option is used

On switch to T.38  we go to T.38<>T.38. If that is not successful and option -F is
not used (fallback not allowed) then we fallback to audio and use T.38<>G.711.

When T.38 is not enabled (-a option used), we use T.38<>G.711 for the entire call.

Only issue is that while in PCM16<>G.711 we do not send any tones. In the receive
case there will be no CED for 1 second. In the send case, there will be no CNG until
the -X n timeout has expired.