Changes to go with Opal fix for switching to T.38.
With a change in Opal we can now switch from T38<>PCM16<>G711 to T38<>T38.
  - Removed offer of PCM16. All calls use T38 at all times.
  - Removed 1 second timer for switch on Receive.

Now all calls start out with:

T38<>PCM16<>G711

During this initial period we will be sending CED or CNG generated by
SpanDSP.

An attempt to switch to T38 can occur for the following reasons:
  - Request from the other side (SIP re-Invite to T38)
  - A tone (CED or CNG) was detected. When calling option -e is needed.
  - A timeout for switching to T38 occurs. This is the -X n option.

The switch will then move us to:

T38<>T38

If this is accepted, then the call proceeds in T38 mode.

If switch to T38 is not we accepted
  If option -F is used, fallback to audio is prohibited and call fails.
  Otherwise we go back to G711:

T38<>PCM16<>G711

The call then proceeds in G711 mode.

If no switch is attempted, the call proceeds in G711 mode.

If the -a option is used then the call stays in G711 mode and no
switching is done or accepted:

T38<>PCM16<>G711